Courses category

+++ lib/features/courses/presentation/widgets/course_details_view_body.dart
@@ -56,7 +56,7 @@ class CourseDetailsViewBody extends StatelessWidget {
         Positioned(
           bottom: 0,
           child: CourseEnrollmentBar(
-            price: course.price,
+            price: course.price.toDouble(),
           ),
         ),
       ],
